Create separate SavingsSummary container, move SearchPanel component to a separate folder, update styles
