+++ LabProject/src/components/search-panel/styles.ts
@@ -1,0 +1,18 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+    root: {
+        display: 'flex',
+        flexDirection: 'row',
+        justifyContent: 'space-between',
+    },
+    inputContainer: {
+        flex: 2,
+    },
+    buttonContainer: {
+        flex: 1,
+        marginLeft: 15,
+    },
+});
+
+export default styles;
